requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose
sqlalchemy